+++ Sources/RichTextKit/Context/RichTextContext+Colors.swift
@@ -3,7 +3,7 @@
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-12-08.
-//  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -20,13 +20,7 @@ public extension RichTextContext {
 
     /// Get the value for a certain color.
     func color(for color: RichTextColor) -> ColorRepresentable? {
-        switch color {
-        case .foreground: foregroundColor
-        case .background: backgroundColor
-        case .strikethrough: strikethroughColor
-        case .stroke: strokeColor
-        case .underline: underlineColor
-        }
+        colors[color]
     }
 
     /// Set the value for a certain color.
@@ -36,13 +30,18 @@ public extension RichTextContext {
     ) {
         guard self.color(for: color) != val else { return }
         userActionPublisher.send(.setColor(color, val))
+        setColorInternal(color, to: val)
+    }
+}
 
-        switch color {
-        case .foreground: foregroundColor = val
-        case .background: backgroundColor = val
-        case .strikethrough: strikethroughColor = val
-        case .stroke: strokeColor = val
-        case .underline: underlineColor = val
-        }
+extension RichTextContext {
+
+    /// Set the value for a certain color, or remove it.
+    func setColorInternal(
+        _ color: RichTextColor,
+        to val: ColorRepresentable?
+    ) {
+        guard let val else { return colors[color] = nil }
+        colors[color] = val
     }
 }