+++ RELEASE_NOTES.md
@@ -50,7 +50,8 @@ By deprecating these functions, we can simplify the library in 1.0, and focus mo
 * `RichTextCommand.IndentOptionsGroup` has been deprecated.
 * `RichTextCommand.StyleOptionsGroup` has been deprecated.
 * `RichTextCommand.SuperscriptOptionsGroup` has been deprecated.
-* `RichTextCoordinator` functions that called `handle(_:)` have been deprecated.
+* `RichTextContext` replaces individual colors with a single `colors`.
+* `RichTextCoordinator` functions that call `handle(_:)` have been deprecated.
 * `RTKL10n.bundle` has been deprecated since we can just use `.module` from now.
 
 ### 💥 Breaking Changes